add role & permission | assgin permission to any module

+++ app/Http/Controllers/ClientController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\Models\Client;
+use Illuminate\Support\Facades\Gate;
+use App\Providers\RouteServiceProvider;
 use Illuminate\Support\Facades\Storage;
 use App\Http\Requests\StoreClientRequest;
 use App\Http\Requests\UpdateClientRequest;
@@ -15,6 +17,9 @@ class ClientController extends Controller
     public function index()
     {
         //
+        if (!Gate::any(['client-anyView', 'client-create', 'client-edit', 'client-view', 'client-delete'])) {
+            return redirect(RouteServiceProvider::HOME);
+        }
         $clients = Client::paginate(config('paginate.count'));
         return view('admin.clients.index', get_defined_vars());
     }
@@ -25,6 +30,7 @@ class ClientController extends Controller
     public function create()
     {
         //
+        Gate::authorize('client-create');
         return view('admin.clients.create');
     }
 
@@ -34,6 +40,7 @@ class ClientController extends Controller
     public function store(StoreClientRequest $request)
     {
         //
+        Gate::authorize('client-create');
         $data = $request->validated();
         //upload image
         $image = $request->image;
@@ -50,6 +57,7 @@ class ClientController extends Controller
     public function show(Client $client)
     {
         //
+        Gate::authorize('client-view');
         return view('admin.clients.show', get_defined_vars());
     }
 
@@ -59,6 +67,7 @@ class ClientController extends Controller
     public function edit(Client $client)
     {
         //
+        Gate::authorize('client-edit');
         return view('admin.clients.edit', get_defined_vars());
     }
 
@@ -68,6 +77,7 @@ class ClientController extends Controller
     public function update(UpdateClientRequest $request, Client $client)
     {
         //
+        Gate::authorize('client-edit');
         $data = $request->validated();
         if ($request->hasFile('image')) {
             Storage::delete("public/member/$client->image");
@@ -86,6 +96,7 @@ class ClientController extends Controller
     public function destroy(Client $client)
     {
         //
+        Gate::authorize('client-delete');
         $client->delete();
         return to_route('admin.clients.index')->with('success', __('admin.delete_successfully'));
     }
