+++ app/Http/Controllers/MessageController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\Models\Message;
+use Illuminate\Support\Facades\Gate;
+use App\Providers\RouteServiceProvider;
 use App\Http\Requests\StoreMessageRequest;
 use App\Http\Requests\UpdateMessageRequest;
 
@@ -14,6 +16,9 @@ class MessageController extends Controller
     public function index()
     {
         //
+        if (!Gate::any(['message-anyView', 'message-create', 'message-edit', 'message-view', 'message-delete'])) {
+            return redirect(RouteServiceProvider::HOME);
+        }
         $messages = Message::paginate(config('paginate.count'));
         return view('admin.messages.index', get_defined_vars());
     }
@@ -24,6 +29,7 @@ class MessageController extends Controller
     public function show(Message $message)
     {
         //
+        Gate::authorize('message-view');
         return view('admin.messages.show', get_defined_vars());
     }
 
@@ -33,6 +39,7 @@ class MessageController extends Controller
     public function destroy(Message $message)
     {
         //
+        Gate::authorize('message-delete');
         $message->delete();
         return to_route('admin.messages.index')->with('success', __('admin.delete_successfully'));
     }
