+++ app/Http/Controllers/ServiceController.php
@@ -3,9 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\Service;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\StoreServiceRequest;
 use App\Http\Requests\UpdateServiceRequest;
-
+use App\Providers\RouteServiceProvider;
 
 class ServiceController extends Controller
 {
@@ -15,7 +16,9 @@ class ServiceController extends Controller
     public function index()
     {
         //
-        
+        if (!Gate::any(['service-anyView', 'service-create', 'service-edit', 'service-view', 'service-delete'])) {
+            return redirect(RouteServiceProvider::HOME);
+        }
         $services = Service::paginate(config('paginate.count'));
         return view('admin.services.index', get_defined_vars());
     }
@@ -26,6 +29,7 @@ class ServiceController extends Controller
     public function create()
     {
         //
+        Gate::authorize('service-create');
         return view("admin.services.create");
     }
 
@@ -35,6 +39,7 @@ class ServiceController extends Controller
     public function store(StoreServiceRequest $request)
     {
         //
+        Gate::authorize('service-create');
         $data = $request->validated();
         Service::create($data);
         return to_route('admin.services.index')->with('success', __('admin.create_successfully'));
@@ -46,6 +51,7 @@ class ServiceController extends Controller
     public function show(Service $service)
     {
         //
+        Gate::authorize('service-view');
         return view('admin.services.show', get_defined_vars());
     }
 
@@ -55,6 +61,7 @@ class ServiceController extends Controller
     public function edit(Service $service)
     {
         //
+        Gate::authorize('service-edit');
         return view('admin.services.edit', get_defined_vars());
     }
 
@@ -64,6 +71,7 @@ class ServiceController extends Controller
     public function update(UpdateServiceRequest $request, Service $service)
     {
         //
+        Gate::authorize('service-edit');
         $data = $request->validated();
         $service->update($data);
         return to_route('admin.services.index')->with('success', __('admin.update_successfully'));
@@ -75,6 +83,7 @@ class ServiceController extends Controller
     public function destroy(Service $service)
     {
         //
+        Gate::authorize('service-delete');
         $service->delete();
         return to_route('admin.services.index')->with('success', __('admin.delete_successfully'));
     }
